requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-anymail==12.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.8
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
